==> paper_code/notebooks/1_checkm_filter.ipynb <==
# 过滤低质量基因组数据

前期已经获得216408条数据，现过滤掉低质量基因组数据

from pathlib import Path

tables1_path = Path("../data/TableS1.csv")
checkm_path = Path("../data/checkm.csv")

import pandas as pd

checkm_df = pd.read_csv(checkm_path)
checkm_df = checkm_df.iloc[:, 1:]
checkm_df.head()

# 筛选完整度≥90%污染率≤5%的基因组
filtered_df = checkm_df[(checkm_df['CheckM completeness'] >= 90) & (checkm_df['CheckM contamination'] <= 5)]
filtered_df.shape

# 读取TableS1
tables1_df = pd.read_csv(tables1_path)
tables1_df.head()

# 将filtered_df的"Assembly Accession"改为"assembly_accession"
filtered_df = filtered_df.rename(columns={"Assembly Accession": "assembly_accession"})
filtered_df.columns

# 提取合并高质量基因组的16S序列和BGC信息
hq_df = pd.merge(left=filtered_df, left_on=["assembly_accession"], right=tables1_df, right_on=["assembly_accession"], how="inner")
hq_df.shape

hq_df.head()

hq_df.to_csv("../data/TableS1_HQ.csv", index=False)
