==> paper_code/notebooks/4_filtered_data_integration.ipynb <==
# 根据过滤后的16S数据更新BGC数据表

from Bio import SeqIO
from pathlib import Path

# 读取fasta文件并获取所有序列名
def get_fasta_sequence_names(fasta_path):
    records = SeqIO.parse(fasta_path, 'fasta')
    sequence_names = [record.id for record in records]
    return sequence_names

fasta_file = Path('../data/filtered_16SrRNA.fasta')
sequence_names = get_fasta_sequence_names(fasta_file)
len(sequence_names)
    

import pandas as pd
tables1_hq_path = Path('../data/tables1_hq.csv')
tables1_hq_df = pd.read_csv(tables1_hq_path)
print(tables1_hq_df.shape)
tables1_hq_df.head()

# 删除16S_ID列中值不在sequence_names里的行
tables1_hq_df_filtered = tables1_hq_df[tables1_hq_df['16S_ID'].isin(sequence_names)]
print(tables1_hq_df_filtered.shape)
tables1_hq_df_filtered.head()

tables1_hq_df_filtered.to_csv("../data/TableS1_HQ_filtered16S.csv", index=False)

# 根据过滤后的 Table 更新 16S 序列

unique_refseq_ids = tables1_hq_df_filtered["16S_ID"].unique()

from Bio import SeqIO
from pathlib import Path

# 读取fasta文件
def read_fasta(fasta_path):
    records = list(SeqIO.parse(fasta_path, 'fasta'))
    return records

# 保留整合后的序列
def filter_not_in_dataset(records):
    keep_records = []
    for record in records:
        if record.id in unique_refseq_ids:
            keep_records.append(record)
    return keep_records


records = read_fasta("../data/filtered_16SrRNA.fasta")
keep_records = filter_not_in_dataset(records)


len(keep_records)

# 保存过滤后的序列
output_file = Path('../data/filtered_intergrated_16SrRNA.fasta')
SeqIO.write(keep_records, output_file, 'fasta')
