==> paper_code/notebooks/7_Fig1_lineage_stat_BGC_stat.ipynb <==
import pandas as pd

dataset_df = pd.read_csv("data/TableS1_HQ_filtered16S.csv")
dataset_df.shape

dataset_df.head()

lineage_df = pd.read_csv("data/psmpa_metadata.tsv", sep='\t')
lineage_df.head()

lineage_info_df = pd.concat([lineage_df[["assembly_accession", "rna_length"]], lineage_df.iloc[:, -7:]], axis=1)
lineage_info_df

merged_df = pd.merge(left=dataset_df, right=lineage_info_df, how="inner", left_on=["assembly_accession"], right_on=["assembly_accession"])
merged_df.head()

merged_df.shape

unique_counts = merged_df.iloc[:, -7:].nunique()
print(unique_counts)


phylum_genus_counts = merged_df.groupby("phylum")["genus"].nunique().sort_values(ascending=False)
print(phylum_genus_counts)

phylum_genus_counts.sum()

# 找出在多个 phylum 下出现的 genus
cross_phylum_genus = merged_df.groupby("genus")["phylum"].nunique()
cross_phylum_genus = cross_phylum_genus[cross_phylum_genus > 1]

# 输出总共有多少个 genus 出现在多个 phylum 中
print(f"有 {len(cross_phylum_genus)} 个 genus 出现在多个 phylum 中：")
print(cross_phylum_genus)

# 查看这些 genus 分别在哪些 phylum 中
for genus in cross_phylum_genus.index:
    phyla = merged_df[merged_df["genus"] == genus]["phylum"].unique()
    print(f"\nGenus: {genus}\nPhyla: {phyla}")


- 对于g__Ponticoccus，确实有2个来源'p__Proteobacteria' 'p__Actinobacteria'；
- 对于g__unclassified Bacteria genus，考虑将'p__unclassified Bacteria phylum' 'p__unclassified Bacteria'合并

merged_df[merged_df['phylum'] == 'p__unclassified Bacteria phylum']

merged_df[merged_df['phylum'] == 'p__unclassified Bacteria']

new_merged_df = merged_df.copy()
new_merged_df.loc[new_merged_df['phylum'] == 'p__unclassified Bacteria phylum', 'phylum'] = 'p__unclassified Bacteria'

new_merged_df[new_merged_df['phylum'] == 'p__unclassified Bacteria phylum']

phylum_genus_counts = new_merged_df.groupby("phylum")["genus"].nunique().sort_values(ascending=False)
phylum_genus_counts.sum()

unique_counts = new_merged_df.iloc[:, -7:].nunique()
print(unique_counts)

unique_counts["genus"] += 1  # g__Ponticoccus 算2个不同类
unique_counts

import pandas as pd

# 创建一个 ExcelWriter 对象
with pd.ExcelWriter("data/Fig1.xlsx") as writer:
    # 将 unique_counts 保存到 sheet1
    unique_counts_df = unique_counts.to_frame(name="unique_counts")  # 转换为 DataFrame
    unique_counts_df.to_excel(writer, sheet_name="sheet1")

    # 将 phylum_genus_counts 保存到 sheet2
    phylum_genus_counts_df = phylum_genus_counts.to_frame(name="phylum_genus_counts")  # 转换为 DataFrame
    phylum_genus_counts_df.to_excel(writer, sheet_name="sheet2")


# 统计每个phylum-genus下的BGC的平均总数
mean_sum_result = new_merged_df.groupby(by=["phylum", "genus"])["Sum"].mean()
mean_sum_result

import numpy as np

# 统计每组的行数
group_sizes = new_merged_df.groupby(by=["phylum", "genus"]).size()
group_sizes.name = "size"
group_sizes

# 对 group_sizes 取 log10
group_sizes_log10 = np.log10(group_sizes)

group_sizes_log10

counts = new_merged_df['16S_ID'].value_counts()
max_count = counts.max()
most_frequent_count = counts[counts == max_count]
most_frequent_count.index[0]

# 计算每个分组中的 "16S_ID" 出现次数
def most_frequent_16S(group):
    # 计算每个 "16S_ID" 的出现次数
    counts = group['16S_ID'].value_counts()
    
    # 获取最大出现次数
    max_count = counts.max()
    
    # 获取所有出现次数为最大值的 16S_ID
    most_frequent_count = counts[counts == max_count]
    
    # 如果只有一个 16S_ID 出现次数最多，返回该 16S_ID；如果有多个，则随机选择一个
    most_frequent = most_frequent_count.sample(n=1, random_state=2).index[0]  # 随机选择一个
    
    return most_frequent

# 统计每个phylum-genus下的代表性16S
most_frequent_16S_result = new_merged_df.groupby(by=["phylum", "genus"]).apply(most_frequent_16S)
most_frequent_16S_result

# 查找重复的值
duplicates = most_frequent_16S_result[most_frequent_16S_result.duplicated(keep=False)]

# 打印重复的值
print(duplicates)
len(duplicates)

# 3. 合并两个结果
merged_result = pd.concat([most_frequent_16S_result, mean_sum_result, group_sizes_log10], axis=1)
merged_result

merged_result.to_csv("data/Fig1C.csv")

target_seq = most_frequent_16S_result.to_list()
target_seq

from Bio import SeqIO
from pathlib import Path

# 读取fasta文件
def read_fasta(fasta_path):
    records = list(SeqIO.parse(fasta_path, 'fasta'))
    return records

# 找出代表性序列用于建树
def get_target_records(records):
    keep_records = []
    for record in records:
        if record.id in target_seq:
            keep_records.append(record)
    return keep_records

records = read_fasta("data/filtered_intergrated_16SrRNA.fasta")
target_records = get_target_records(records)

# 保存序列
output_file = Path('data/Fig1C_genus_rep_16SrRNA.fasta')
SeqIO.write(target_records, output_file, 'fasta')


new_merged_df.to_csv("data/TableS1_HQ_filtered16S_lineage.csv", index=False)
